pr feedback: move utility files
