+++ Assets/MixedRealityToolkit/_Core/Utilities/EdgeUtilities.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities;
+using Microsoft.MixedReality.Toolkit.Internal.Definitions.BoundarySystem;
 using UnityEngine;
 
 namespace Microsoft.MixedReality.Toolkit.Internal.Utilities
